#36 feat: WebRTC 공부방 페이지에 적용
