+++ backend/package.json
@@ -30,6 +30,8 @@
     "@nestjs/platform-ws": "^9.2.0",
     "@nestjs/swagger": "^6.1.3",
     "@nestjs/typeorm": "^9.0.1",
+    "cache-manager": "^5.1.3",
+    "cache-manager-ioredis": "^2.1.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.13.2",
     "cross-env": "^7.0.3",
